requirements.txt
polars
pandas
xlsxwriter
python-dotenv
pywin32
sqlalchemy
fastapi
uvicorn
openai
pydantic
requests